refactor(attendances): separate getUserAttendances into controller and model
